+++ src/civilization/population_expansion/population_expansion_components.rs
